requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
idna==3.4
Naked==0.1.31
pyasn1==0.4.8
pycryptodome==3.15.0
Pygments==2.13.0
PyYAML==6.0
requests==2.28.1
rich==12.6.0
rsa==4.9
shellescape==3.8.1
shellingham==1.5.0
typer==0.6.1
typing-extensions==4.4.0
urllib3==1.26.12
